requirements.txt
appdirs==1.4.2
builtwith==1.3.2
future==0.16.0
packaging==16.8
pyparsing==2.1.10
python-whois==0.6.5
six==1.10.0
